% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{annotateGenes}
\alias{annotateGenes}
\title{Annotate gene using protein atlas}
\usage{
annotateGenes(genes, ...)
}
\arguments{
\item{genes}{A vector of genes}

\item{...}{other arguments}
}
\value{
A vector of annotated categories
}
\description{
Annotate gene using protein atlas
}
