% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{apply_MM}
\alias{apply_MM}
\title{Apply for Sparse matrix}
\usage{
apply_MM(X, MARGIN = 1, FUN)
}
\arguments{
\item{X, }{a sparse matrix}

\item{MARGIN}{1 or 2, indicating row-wise or column-wise}

\item{FUN}{a function}
}
\value{
Numeric vector
}
\description{
Use it with cautions! Only apply FUN to all the non-zero values!!
}
