% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{batch_testing}
\alias{batch_testing}
\title{Wrapper for batch testing}
\usage{
batch_testing(
  data_df,
  group_var,
  test_var,
  test_set,
  by_facet = NULL,
  method = "permutation",
  pairby = NULL,
  p_cutoff = 0.1,
  falltoperm = F,
  ...
)
}
\arguments{
\item{data_df}{The data frame to test}

\item{group_var}{The variable name for testing groups}

\item{test_var}{The variable name for testing values}

\item{test_set}{A 2 x n combn matrix for pairs of condition to test}

\item{by_facet}{A string for the facet}

\item{method}{One of the c('permutation', <to be added>)}

\item{pairby}{The variable name for pairing, default NULL}

\item{p_cutoff}{Specify the cutoff for significant p values}

\item{...}{Other parameters for test functions}
}
\value{
A data.frame
}
\description{
Wrapper for batch testing
}
