% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dea.wilcox}
\alias{dea.wilcox}
\title{Differential expression analysis using Wilcoxon test}
\usage{
dea.wilcox(x, ...)
}
\arguments{
\item{x}{A list with two named expression matrixces of n genes * m cells.}

\item{...}{Other options. Placeholder, not used yet.}
}
\value{
a data.frame
}
\description{
Differential expression analysis using Wilcoxon test
}
