% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dirichletTest}
\alias{dirichletTest}
\title{Dirichlet-multinomial regression test}
\usage{
dirichletTest(count_df, predictor)
}
\arguments{
\item{count_df}{sample x cell_fraction data.frame}

\item{predictor}{A named vector}
}
\description{
Dirichlet-multinomial regression test
}
