% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{donuts}
\alias{donuts}
\title{x      numeric vector for each slice
group  vector identifying the group for each slice
labels vector of labels for individual slices
col    colors for each group
radius radius for inner and outer pie (usually in [0,1])}
\usage{
donuts(x, group = 1, labels = NA, col = NULL, radius = c(0.7, 1))
}
\description{
x      numeric vector for each slice
group  vector identifying the group for each slice
labels vector of labels for individual slices
col    colors for each group
radius radius for inner and outer pie (usually in [0,1])
}
