% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{expand.mt}
\alias{expand.mt}
\title{Expand matrix}
\usage{
\method{expand}{mt}(x_mt, col_names = NULL, row_names = NULL, fill = 0)
}
\arguments{
\item{x_mt}{a matrix}

\item{col_names}{target colnames}

\item{row_names}{target rownames}
}
\value{
a expanded matrix
}
\description{
Expand matrix
}
