% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{expand.vec}
\alias{expand.vec}
\title{Expand vector}
\usage{
\method{expand}{vec}(x_vec, vec_names = NULL, fill = 0)
}
\arguments{
\item{x_vec}{original vector}

\item{vec_names}{target vector names}
}
\value{
a expanded vector
}
\description{
Expand vector
}
