% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{extractLegend}
\alias{extractLegend}
\title{Extract legend from a ggplot object}
\usage{
extractLegend(a.gplot)
}
\arguments{
\item{a.gplot}{A ggplot object}
}
\description{
Extract legend from a ggplot object
}
