% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{EXWB.initiate}
\alias{EXWB.initiate}
\title{Initiate a openxlsx workbook}
\usage{
EXWB.initiate(...)
}
\arguments{
\item{...}{Need nothing}
}
\description{
Inititate a openxlsx workbook with hard-coded format. Will be upgraded in the future (if necessary)
}
