% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{EXWB.writeSheet}
\alias{EXWB.writeSheet}
\title{Initiate a openxlsx workbook}
\usage{
EXWB.writeSheet(wb, sheetName, sheetData, overwrite = F, ...)
}
\arguments{
\item{wb}{An openxlsx wb object to be written in}

\item{sheetName}{String; Name of the sheet to write.
Sheet name must have <= 31 characters without any strange characters like !\@\#\$\%\^\&\*\(\)}

\item{sheetData}{Data frame or tibble to write.}

\item{overwrite}{Logical; if to overwrite the sheet with the same name.}
}
\description{
Inititate a openxlsx workbook with hard-coded format. Will be upgraded in the future (if necessary)
}
