% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{fisherTest}
\alias{fisherTest}
\title{Batch fisher test}
\usage{
fisherTest(counts_bystatus, p.adjust.method = "BH", ...)
}
\arguments{
\item{counts_bystatus}{A x B count matrix}

\item{p.adjust.method}{P adjust method for `p.adjust`, default = 'BH'}

\item{...}{other arguments for `fisher.test`}
}
\value{
a data.frame for p values and ORs from `fisher.test`
}
\description{
Batch fisher test
}
