% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Generate ggplot default colors}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{integer, number of the generated colors}
}
\description{
adopted from \url{https://stackoverflow.com/questions/8197559/emulate-ggplot2-default-color-palette}
}
