% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{loginfo}
\alias{loginfo}
\title{Print formated log}
\usage{
loginfo(..., printnow = T)
}
\arguments{
\item{...}{a string or list of strings to print}

\item{printnow}{logical, if print on the screen or return log as a string}
}
\description{
Print formated log
}
