% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{msgASAP}
\alias{msgASAP}
\title{Sent message to wechat}
\usage{
msgASAP(ttl, msg = "", SCKEY)
}
\arguments{
\item{ttl}{string, title of the message}

\item{msg}{string, optional}

\item{SCKEY}{string, SCKEY from ServerChan, must be provided}
}
\description{
Sent message to wechat
}
