% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{processBar}
\alias{processBar}
\title{Process bar for loops}
\usage{
processBar(
  objName,
  i,
  cycles,
  title = "Process",
  scale = 40,
  sign = "#",
  tail = "",
  terminal = "R",
  final = "Work done!"
)
}
\description{
a terminal process Bar, use initiatePB to initialize at the beginning of the forloop,
and insert processBar with objName, i, cycles inside the for loop
}
